+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Improvement to Enter auto-indent when text after cursor has closing braces.
  * Minimum and recommended CDT version is now `8.8`.
  * When debugging, fixed toggling breakpoints on and off for files that are outside the workspace.
  * When debugging, fixed opening source files that are are outside the workspace.